Moved loss.py to robust_loss.py
